// backend/src/interfaces/familyUpdate.ts
interface familyUpdate {
  data: string;
  categoria: string;
  nome: string;
  contratante: string;
  local?: string; 
  primeiroBackupBruto?: number;
  primeiroBackupBrutoTamanho?: number;
  segundoBackupBruto?: number;
  segundoBackupBrutoTamanho?: number;
  primeiroBackup?: number;
  primeiroBackupTamanho?: number;
  segundoBackup?: number;
  segundoBackupTamanho?: number;
}

export default familyUpdate;